much better popup placement, all css
incorporates @ojame's pull request fixing #7. Removes all DOM
calculation stuff on the pop placement and makes attaches all the work
on the rw-popup-container class.  Adds a new less var to adjust the
popup left/right border offsets of the parent.

also fixes an annoying bug in IE11 that was maintaining permanent focus
on the select