Leverage parser for passing strings as date value
This makes it easier to pass in strings as the date `value` and `defaultValue` instead of limiting these to a date object. This is useful since passing in a string from the model is likely the most common use-case.

A new prop called `valueFormat` is also exposed to specify the format for the `value` and `defaultValue`. Although the `editFormat` and `format` props are also tried for parsing the date string making `defaultValue` optional.