Preserve text selection on tab
Standard web browser behavior is to select all text within a text input
element when tabbing to that element.  NumberPicker behaves this way by
default; _however_, if a format is set, so the string value that's
displayed while the input has the focus is different than the formatted
value that's displayed normally, then switching from the formatted value
to the display value loses the text selection.

The only idea I had for a solution was to check if all text is selected
before update and, if that's the case, restore the selection after
update.

Fixes #375.