Add onDelete callback prop to Multiselect, called on tag item deletion
The callback is given two arguments: the tag item and the event from the "cross"
delete-item button. This allows the client code, among other things, to prevent event
propagation to the Multiselect control itself. This can be useful for e.g. preventing
the drop down list from opening when a tag item is deleted.